fix(assets): remove duplicate Mahjong tile asset load on web (#1629)
Both GameCanvas.web.tsx and MahjongScreen.tsx independently called
Asset.fromModule().downloadAsync() over all 42 SVG tiles on mount,
causing 84 network requests instead of 42.

Introduce tileAssetLoader.web.ts with a module-level singleton promise
so the first caller downloads all tiles and every subsequent caller
(including concurrent ones) awaits the same in-flight result. A native
stub (tileAssetLoader.ts) is provided for Metro platform resolution.

GameCanvas.web.tsx now converts the pre-downloaded URIs into
HTMLImageElement objects; MahjongScreen.tsx reads the same URIs for the
FlyingPair overlay. expo-asset is touched exactly once per session.

Closes #1629
